Factor out RPC consoles ParseCommandLine. Add callbacks for flexibility
